+++ src/tests/integration/TestFiles/JsonOptions.cs
@@ -1,4 +1,7 @@
-﻿using Seekatar.OptionToStringGenerator;
+﻿#nullable enable
+using System;
+using System.IO;
+using Seekatar.OptionToStringGenerator;
 
 namespace Test;