+++ src/tests/integration/TestFiles/ObjectMasking.cs
@@ -1,6 +1,12 @@
 ﻿namespace Test;
 using Seekatar.OptionToStringGenerator;
 
+class OmClass
+{
+    public string Name { get; set; } = string.Empty;
+    public override string ToString() => $"{nameof(OmClass)}: {Name}";
+}
+
 [OptionsToString]
 class ObjectMasking
 {
@@ -10,5 +16,5 @@ class ObjectMasking
     }
 
     [OutputRegex(Regex = "User Id=([^;]+).*Password=([^;]+)", IgnoreCase = true)]
-    public PublicOptions.AClass AnObject { get; }
+    public OmClass AnObject { get; }
 }