+++ src/tests/integration/TestFiles/TripleNestedOptions.cs
@@ -1,4 +1,5 @@
-﻿namespace Test;
+﻿#nullable enable
+namespace Test;
 
 using Seekatar.OptionToStringGenerator;
 using System.ComponentModel.DataAnnotations;